temporal_worker/wf_mgt_api/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
temporalio==1.6.0
psycopg[binary,pool]==3.3.3
pydantic==2.12.5
python-multipart==0.0.22
azure.storage.blob==12.28.0
azure-ai-formrecognizer==3.3.0
python-dotenv==1.2.2
requests==2.32.5
aiohttp==3.13.3
openai==1.99.6